pi/QueingSystem/code/picode/requirements.txt
flask==2.0.1
flask-cors==3.0.10
firebase-admin==5.0.3
python-dotenv==0.19.0
requests==2.26.0
opencv-python==4.8.1.78
Pillow==10.1.0
pygame==2.5.2 